ci(consensus): split flaky scenario tests into a non-blocking lane
The consensus integration suite has a clean stable/flaky split under
stress: a small core (cluster forms + replicates, single-node produce +
finalize, raft-history clear, quorum-loss recovery) never failed in a
20x baseline run, while the leader-change / restart / quorum-loss /
tx-forwarding scenarios flake at ~5-25%. That flakiness is a
test-environment artifact — reth's session manager mishandles the
simultaneous cross-dial that localhost + synchronized startup makes
near-certain (a dup-dial that staggered, real-RTT production deploys
almost never hit) — not a demonstrated production bug.

Rather than retry failed tests (hides real regressions) or keep the
whole suite flaky, move the flaky scenarios into a
`consensus_node::nonblocking` module and run them in a dedicated CI job
that is NOT in the required-checks list. They still run on every PR for
visibility; they just don't randomly block unrelated merges. The four
stable tests stay in the required `test` job and get stress-validated
(the branch's --stress-count still applies there).

- mod.rs keeps the 4 stable tests + shared helpers (three made
  pub(crate) for the new module).
- nonblocking/ holds the 8 moved tests + the restarted_node_catchup
  submodule (all 4 of its tests are restart-based → non-blocking).
- ci.yml: `test` job excludes `consensus_node::nonblocking`; new
  non-required `consensus-nonblocking` job runs exactly that lane once.

The deeper fix (graceful leader demotion so churn from any cause —
flap, restart, quorum blip, probe timeout — stops detonating a node)
is captured as a follow-up; it's the only cause-agnostic fix but is a
real change to core block production and isn't justified by an unproven
production problem.
